issue #84: extracted location from stop and city entity
