+++ java/org.travlyn.shared/src/main/java/org/travlyn/shared/model/db/CityEntity.java
@@ -9,82 +9,13 @@ import java.util.Set;
 
 @Entity
 @Table(name = "city")
-public class CityEntity implements DataEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private int id;
-
-    @Column(name = "longitude")
-    private double longitude;
-
-    @Column(name = "latitude")
-    private double latitude;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "image")
-    private String image;
-
-    @Column(name = "description", length = 65536)
-    private String description;
+public class CityEntity extends Location{
 
     @Column(name = "unfetched_stops")
     private boolean unfetchedStops;
 
     @OneToMany(orphanRemoval = true, cascade = {CascadeType.PERSIST, CascadeType.MERGE})
     private Set<StopEntity> stops = new HashSet<>();
-
-    @Override
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public double getLongitude() {
-        return longitude;
-    }
-
-    public void setLongitude(double longitude) {
-        this.longitude = longitude;
-    }
-
-    public double getLatitude() {
-        return latitude;
-    }
-
-    public void setLatitude(double latitude) {
-        this.latitude = latitude;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
 
     public Set<StopEntity> getStops() {
         return stops;
@@ -105,12 +36,12 @@ public class CityEntity implements DataEntity {
     @Override
     public City toDataTransferObject() {
         City city = new City()
-                .id(this.id)
-                .longitude(this.longitude)
-                .latitude(this.latitude)
-                .name(this.name)
-                .image(this.image)
-                .description(this.description)
+                .id(super.getId())
+                .longitude(super.getLongitude())
+                .latitude(super.getLatitude())
+                .name(super.getName())
+                .image(super.getImage())
+                .description(super.getDescription())
                 .setUnfetchedStops(this.unfetchedStops);
         HashSet<Stop> stopHashSet = new HashSet<>();
         for (StopEntity stopEntity : stops) {
